requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv
slowapi
pydantic
beautifulsoup4